General code cleanup and edited mockup
