--- src/App.css
@@ -1,17 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.center {
-  display: flex;
-  justify-content: center;
-}
-
-.particles {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: -1;
-}
