+++ src/components/ImageLinkForm/ImageLinkForm.js
@@ -14,11 +14,14 @@ const ImageLinkForm = ({onInputChange, onSubmit}) => {
                     <input 
                         className="f4 pa2 w-70 center" 
                         type="text" 
-                        onChange={onInputChange} />
+                        onChange={onInputChange}
+                        required />
                     <button 
                         className="w-30 grow f4 link ph3 pv2 dib white 
                             bg-light-purple" 
-                        onClick={onSubmit}>Detect</button>
+                        onClick={onSubmit}>
+                        Detect
+                    </button>
                 </div>
             </div>
         </div>
