+++ src/components/Splash/Splash.js
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const Splash = ({ onRouteChange }) => {
+    return (
+        <div className='white f3'>
+            <h1>Welcome to MugFinder!</h1>
+            <p>You give us the URL of a photo, 
+                we’ll show you any faces in it.
+            </p>
+            <img src="" />
+            <p>
+                <p>Create an account</p> to get started.
+            </p>
+        </div>
+    );
+}
+
+export default Rank;
